+++ src/physics/fallingsand/chunks.rs
@@ -2,3 +2,4 @@ mod chunk;
 mod core;
 mod partial_layer;
 pub mod radial_mesh;
+pub mod util;
