+++ src/physics/fallingsand/chunks/chunk.rs
@@ -1,12 +1,8 @@
 use ggez::glam::Vec2;
 use ggez::graphics::Image;
+use ggez::graphics::Rect;
 use ggez::graphics::Vertex;
 use ggez::Context;
-
-/// Finds a point halfway between two points
-pub fn interpolate_points(p1: &Vec2, p2: &Vec2) -> Vec2 {
-    Vec2::new((p1.x + p2.x) * 0.5, (p1.y + p2.y) * 0.5)
-}
 
 /// A chunk that can be rendered and simulated
 pub trait Chunk {
@@ -39,6 +35,7 @@ pub trait Chunk {
     fn get_end_radial_theta(&self) -> f32;
     fn get_num_radial_lines(&self) -> usize;
     fn get_num_concentric_circles(&self) -> usize;
+    fn get_bounding_box(&self) -> Rect;
     fn total_size(&self) -> usize {
         self.get_num_radial_lines() * self.get_num_concentric_circles()
     }
